Update emulator binaries to fix platform boot.
This updates the emulator binaries to the latest tip-of-tree
from the studio-1.3-dev, mostly to fix the boot of the AOSP system
image, by using the patch from:

    https://android-review.googlesource.com/#/c/143632/

Note that the boot is still broken for other reasons, but this is
a necessary step to properly diagnose it.

See README file for the full change log.

Change-Id: I4f27e11a44bdca223ec4bdc2dab57444673ecbac